test-requirements.txt
molecule
molecule-docker
docker
ansible-lint==5.0.1
yamllint
flake8
jmespath
pytest
pytest-testinfra
tox
tox-gh-actions

git+https://github.com/bodsch/pyjolokia@0.3.2
